moved Tournament Search page to it's own directory
